backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
pillow-heif==0.13.1
numpy==1.26.2
python-magic==0.4.27
aiofiles==23.2.1
httpx==0.25.2
watchdog==3.0.0